if, else and while task 5

+++ variables_if_else_while/5-print_numbers.c
@@ -5,7 +5,7 @@
 * Return: 0 if exited correctly, non-zero otherwise.
 */
 
-long main(void)
+int main(void)
 {
 	long l = 0123456789
